Changes For Gem
These changes were suggested by Drew DePonte to make everything more
gem-like.
