--- spec/lib/dot_grid/dot_grid_generator_spec.rb
@@ -1,4 +1,0 @@
-require 'spec_helper'
-
-describe "DotGridGenerator" do
-end
