+++ spec/lib/dot_grid/generator_spec.rb
@@ -1,0 +1,4 @@
+require 'spec_helper'
+
+describe "DotGrid::Generator" do
+end
